Refactor: Move remaining creator components to features directory structure - Moved CreatorInfo to features/creator/info - Moved CreatorTags to features/creator/tags - Moved NotableClients to features/creator/clients - Moved PortfolioGallery to features/creator/portfolio/gallery - Updated main barrel files to export all components - Maintained backward compatibility in original files with deprecated notices - Updated COMPONENT-MOVEMENT-PLAN.md to track progress
