+++ COMPONENT-MOVEMENT-PLAN.md
@@ -49,10 +49,10 @@ This document outlines the plan for moving components to the new directory struc
 | CreatorRating.tsx | /src/components/creator/CreatorRating.tsx | /src/components/features/creator/rating/CreatorRating.tsx | ✅ Completed | Verify ratings display correctly in creator cards |
 | PortfolioPreview.tsx | /src/components/creator/PortfolioPreview.tsx | /src/components/features/creator/portfolio/PortfolioPreview.tsx | ✅ Completed | Verify portfolio previews display in creator cards |
 | types.ts | /src/components/creator/types.ts | /src/components/features/creator/types/index.ts | ✅ Completed | Verify types are correctly exported and used |
-| CreatorInfo.tsx | /src/components/creator/CreatorInfo.tsx | /src/components/features/creator/info/CreatorInfo.tsx | Not Started | Verify creator info displays correctly |
-| NotableClients.tsx | /src/components/creator/NotableClients.tsx | /src/components/features/creator/clients/NotableClients.tsx | Not Started | Verify notable clients section displays correctly |
-| CreatorTags.tsx | /src/components/creator/CreatorTags.tsx | /src/components/features/creator/tags/CreatorTags.tsx | Not Started | Verify creator tags display correctly |
-| PortfolioGallery.tsx | /src/components/creator/PortfolioGallery.tsx | /src/components/features/creator/portfolio/PortfolioGallery.tsx | Not Started | Verify portfolio gallery displays correctly |
+| CreatorInfo.tsx | /src/components/creator/CreatorInfo.tsx | /src/components/features/creator/info/CreatorInfo.tsx | ✅ Completed | Verify creator info displays correctly |
+| NotableClients.tsx | /src/components/creator/NotableClients.tsx | /src/components/features/creator/clients/NotableClients.tsx | ✅ Completed | Verify notable clients section displays correctly |
+| CreatorTags.tsx | /src/components/creator/CreatorTags.tsx | /src/components/features/creator/tags/CreatorTags.tsx | ✅ Completed | Verify creator tags display correctly |
+| PortfolioGallery.tsx | /src/components/creator/PortfolioGallery.tsx | /src/components/features/creator/portfolio/gallery/PortfolioGallery.tsx | ✅ Completed | Verify portfolio gallery displays correctly |
 
 ## Implementation Notes
