+++ src/components/features/creator/clients/README.md
@@ -1,0 +1,14 @@
+# Creator Clients Components
+
+Components for displaying creator's notable clients and related information.
+
+## Components
+
+- `NotableClients.tsx` - Displays a row of client logos with tooltips
+- `index.ts` - Barrel file for clean exports
+
+## Usage
+
+```tsx
+import { NotableClients } from '@/components/features/creator';
+```
