+++ src/components/features/creator/clients/index.ts
@@ -1,0 +1,1 @@
+export { NotableClients } from './NotableClients';
