+++ src/components/features/creator/index.ts
@@ -3,3 +3,6 @@ export * from './media';
 export * from './rating';
 export * from './portfolio';
 export * from './types';
+export * from './info';
+export * from './tags';
+export * from './clients';
