+++ src/components/features/creator/info/README.md
@@ -1,0 +1,14 @@
+# Creator Info Components
+
+Components for displaying creator information like name, location, and services.
+
+## Components
+
+- `CreatorInfo.tsx` - Displays creator name, verification badge, location, and services
+- `index.ts` - Barrel file for clean exports
+
+## Usage
+
+```tsx
+import { CreatorInfo } from '@/components/features/creator';
+```
