+++ src/components/features/creator/info/index.ts
@@ -1,0 +1,1 @@
+export { CreatorInfo } from './CreatorInfo';
