+++ src/components/features/creator/portfolio/gallery/README.md
@@ -1,0 +1,14 @@
+# Creator Portfolio Gallery
+
+Components for displaying a gallery view of creator portfolio work.
+
+## Components
+
+- `PortfolioGallery.tsx` - Grid display of portfolio images with lazy loading
+- `index.ts` - Barrel file for clean exports
+
+## Usage
+
+```tsx
+import { PortfolioGallery } from '@/components/features/creator/portfolio/gallery';
+```
