+++ src/components/features/creator/portfolio/gallery/index.ts
@@ -1,0 +1,1 @@
+export { PortfolioGallery } from './PortfolioGallery';
