+++ src/components/features/creator/portfolio/index.ts
@@ -1,1 +1,2 @@
 export { PortfolioPreview } from './PortfolioPreview';
+export { PortfolioGallery } from './gallery';
