+++ src/components/features/creator/tags/README.md
@@ -1,0 +1,19 @@
+# Creator Tags Components
+
+Components for displaying creator tags/specialties and related styling utilities.
+
+## Components
+
+- `CreatorTags.tsx` - Displays a horizontal scrollable list of creator specialties/tags
+- `index.ts` - Barrel file for clean exports
+
+## Utilities
+
+- `getDefaultTags` - Returns a set of default tags based on creator name and services
+- `getTagStyle` - Returns styling classes for a specific tag category
+
+## Usage
+
+```tsx
+import { CreatorTags, getDefaultTags } from '@/components/features/creator';
+```
