+++ src/components/features/creator/tags/index.ts
@@ -1,0 +1,1 @@
+export { CreatorTags, getDefaultTags, getTagStyle } from './CreatorTags';
